Copter: move update_sensor_flags to GCS_Copter.cpp
